backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
httpx==0.27.2
anthropic==0.39.0
Pillow==10.4.0
pytest==8.3.3
pytest-asyncio==0.24.0
